Feature ETP-1234: Fix boolean selector value change.
