+++ packages/MainUI/app/window/[windowId]/[tabId]/[recordId]/page.tsx
@@ -13,6 +13,7 @@ import { FormProvider, useForm } from 'react-hook-form';
 import { BaseSelector } from '@/components/Form/FormView/selectors/BaseSelector';
 import { useMemo } from 'react';
 import { buildUpdatedValues, getCombinedEntries } from '@/utils';
+import FormView from '@/components/Form/FormView';
 
 function Page({ window, tab }: { window: WindowMetadata; tab: Tab }) {
   const { t } = useTranslation();
@@ -50,18 +51,7 @@ function Page({ window, tab }: { window: WindowMetadata; tab: Tab }) {
     return <Spinner />;
   }
 
-  return (
-    <FormProvider {...form}>
-      <div style={styles.box}>
-        <Toolbar windowId={window.id} tabId={tab.id} isFormView={true} />
-      </div>
-      <div className="p-2 space-y-4">
-        {Object.entries(tab.fields).map(([hqlName, field]) => (
-          <BaseSelector field={field} key={hqlName} />
-        ))}
-      </div>
-    </FormProvider>
-  );
+  return <FormView form={form} tab={tab} window={window} />;
 }
 
 export default function EditRecordPage() {
