--- packages/MainUI/app/window/[windowId]/[tabId]/[recordId]/styles.ts
@@ -1,3 +1,0 @@
-export const styles = {
-  box: { margin: '0 0 0.5rem 0' },
-};
